# sync: eliminate global Mutex in (*Pool).pinSlow operations